services/auth-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
httpx==0.28.1
aiokafka==0.12.0
